requirements.txt
actions==1.0.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
bunch==1.0.1
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
colorama==0.4.6
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.2.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.4
requests==2.31.0
selenium==4.11.2
seletools==1.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
wsproto==1.2.0
zipp==3.15.0
